clean up and unify did code
